chore(scripts): move calibrate to scripts
