+++ pyproject.toml
@@ -162,7 +162,7 @@ all = [
 ]
 
 [project.scripts]
-lerobot-calibrate="lerobot.calibrate:main"
+lerobot-calibrate="lerobot.scripts.lerobot_calibrate:main"
 lerobot-find-cameras="lerobot.scripts.lerobot_find_cameras:main"
 lerobot-find-port="lerobot.scripts.lerobot_find_port:main"
 lerobot-record="lerobot.record:main"
