+++ tests/test_control_robot.py
@@ -16,9 +16,9 @@
 
 from unittest.mock import patch
 
-from lerobot.calibrate import CalibrateConfig, calibrate
 from lerobot.record import DatasetRecordConfig, RecordConfig, record
 from lerobot.replay import DatasetReplayConfig, ReplayConfig, replay
+from lerobot.scripts.lerobot_calibrate import CalibrateConfig, calibrate
 from lerobot.teleoperate import TeleoperateConfig, teleoperate
 from tests.fixtures.constants import DUMMY_REPO_ID
 from tests.mocks.mock_robot import MockRobotConfig
